refactor: command palette file structure
