+++ web/components/command-palette/actions/index.ts
@@ -1,0 +1,5 @@
+export * from "./issue-actions";
+export * from "./help-actions";
+export * from "./project-actions";
+export * from "./theme-actions";
+export * from "./workspace-settings-actions";
