+++ web/components/command-palette/actions/issue-actions/index.ts
@@ -1,0 +1,4 @@
+export * from "./actions-list";
+export * from "./change-state";
+export * from "./change-priority";
+export * from "./change-assignee";
