--- web/components/command-palette/issue/index.ts
@@ -1,3 +1,0 @@
-export * from "./change-issue-state";
-export * from "./change-issue-priority";
-export * from "./change-issue-assignee";
